research-study-platform/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
openai==1.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
xlsxwriter==3.1.9
openpyxl==3.1.2
requests==2.31.0